add MudBlazor
proved that we can have an AppDT with a different version of MudBlazor than what the shell is on.  It still is based on signatures but it is possible.  Added PreserverCompilationContext in project files to try to not

tested dependency injection at Native and Shell layer... stuck on App layer since the plugins are loaded after everything starts up.  The other headache to handle is having multiple versions of the same DLL.  AssemblyLoadContext? Next commit???

Added HybridCoupling project for integrations between Native, Shell and App layers